Refactor codebank into a django app
