+++ mysite/settings.py
@@ -32,7 +32,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'codedoor'
+    'codedoor',
+    'codebank'
 ]
 
 MIDDLEWARE = [
